+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+
+## 0.1.0-alpha.9 (2024-01-23)
+
+- Improve the handling error in `g-composer` and the SVG Template Engine.
+- Added `g-object` directive from Object Plugin.
+- Fixed error in Firefox when the SVG and `g-composer` don't have `width` or `height` values.
+
 ## 0.1.0-alpha.8 (2024-01-12)
 
 - Fixed error loading plugin from `g-composer`.